subd: skip stopping and starting services if update will reboot.
If services are stopped, started and shortly after the system reboots,
this can leave insufficient time for the service to start up before
being shut down for the reboot, thus it might not have time to shut
down cleanly. Since the system is going to reboot anyway, skip the
stop and start actions and just reboot.