lib/slavedriver: bugfix: consumer could wait even when a slave is ready.
Fix a race condition between the roll call (manager) goroutine creating
a slave and a consumer doing an on-demand create (because the idle list
was empty) could lead to the consumer waiting for two create times:
the manager create and the on-demand create.

Solved this and other potential problems by moving all state management
logic into the roll call goroutine and having consumers communicate
with the manager over channels.