hypervisor/manager: fix spurious deadlock when deleting VMs.
The vmInfoType.delete() method grabbed the Manager lock while the VM lock
was held, which violated the normal pattern of grabbing the Manager lock
before grabbing the lock for VM(s). This manifested as deadlocks every
couple of weeks or so with regular VM create/delete activity (~32/hour,
some back-to-back).