imageserver/scanner: remove string de-duplication as it has limited value.
De-duplicating strings in the images saves ~5-8% memory (after garbage
collection). The price to pay for these savings is added complexity and
stalls whenever an image is deleted/expired, as the de-duplication state
has to be rebuilt from all images.

It would be possible to rework the de-duplication by splitting the rebuild
one image at a time, making the reasonable approximation that a missed
image is harmless. However, even then, a more serious problem remains:
the de-duplication happens in the background, after some other goroutine
potentially has a reference to the image. This is a potential race
condition, as string references may be updated while being used.

On balance, the memory optimisation does not seem worth the price, so this
idea is being retired.